+++ gemp-module/gemp-stccg-logic/src/main/java/com/gempukku/stccg/modifiers/Modifier.java
@@ -30,7 +30,7 @@ public interface Modifier {
 
     boolean hasRemovedText(DefaultGame game, PhysicalCard physicalCard);
 
-    int getAttributeModifier(DefaultGame cardGame, PhysicalCard physicalCard);
+    float getAttributeModifier(DefaultGame cardGame, PhysicalCard physicalCard);
 
     boolean cancelsStrengthBonusModifier(DefaultGame game, PhysicalCard modifierSource, PhysicalCard modifierTaget);
